requirements.txt
fastapi
pydantic
uvicorn
#nest-asyncio
requests
simpleeval
pyyaml
sseclient-py
